requirements.txt
python-telegram-bot==20.7
pygithub==2.1.1
python-dotenv==1.0.0
requests==2.31.0
